python/requirements.txt
BeautifulSoup==3.2.1
Cerberus==0.6
#Eve==0.4-dev
# Temporary fix l v0.4 is released
git+https://github.com/nicolaiarocci/eve.git#egg=Eve
Events==0.2.0
Flask==0.10.1
Flask-PyMongo==0.3.0
Jinja2==2.7.2
MarkupSafe==0.21
Werkzeug==0.9.4
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.17
celery==3.1.11
itsdangerous==0.22
kombu==3.0.15
pymongo
pytz==2014.2
redis==2.9.1
requests==2.2.1
simplejson==3.3.0
urllib3==1.8.2
wsgiref==0.1.2
bcrypt
python-oauth2
pocket
gunicorn
sendgrid
flask-cors

